update local jvm with comments
