+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/example/lenetLocal/README.md
@@ -1,7 +1,9 @@
-# Running LeNet5 Model on local JVM
+# Running LeNet5 Model as a local Java/Scala program
 
 This example shows how to run training, prediction and testing with LeNet5  model on local JVM with BigDL. Lenet5 is a classical CNN model used in digital number classification. For detail information,
 please refer to <http://yann.lecun.com/exdb/lenet/>.
+
+To run the BigDL model as a local Java/Scala program without Spark, user needs to set Java property `bigdl.localMode` to `true`. If user wants to specify how many cores to be used for training/testing/prediction, he needs to set Java property `bigdl.coreNumber` to the core number. User can either call `System.setProperty("bigdl.localMode", "true")` and `System.setProperty("bigdl.coreNumber", core_number)` in the Java/Scala code, or passing -Dbigdl.localMode=true and -Dbigdl.coreNumber=core_number when runing the program. In this example, we use the former way to set these Java properties. 
 
 ## Prepare MNIST Data
 You can download the MNIST Data from [here](http://yann.lecun.com/exdb/mnist/). Unzip all the
@@ -19,21 +21,23 @@ You can build one by refer to the
 ## Train the Model
 Example command
 ```
-java -cp dist/lib/bigdl-VERSION-jar-with-dependencies.jar \
+java -cp spark/dl/target/bigdl-VERSION-jar-with-dependencies-and-spark.jar \
 com.intel.analytics.bigdl.example.lenetLocal.Train \
 -f path_to_mnist_folder \
+-c core_number \
 -b batch_size \
 --checkpoint ./model
 ```
 
 In the above commands
 * -f: where you put your MNIST data
+* -c: The core number on local machine used for this training. The default value is physical cores number. Get it through Runtime.getRuntime().availableProcessors() / 2
+* -b: The mini-batch size. It is expected that the mini-batch size is a multiple of core_number
 * --checkpoint: Where you cache the model/train_state snapshot. You should input a folder and
 make sure the folder is created when you run this example. The model snapshot will be named as
 model.#iteration_number, and train state will be named as state.#iteration_number. Note that if
 there are some files already exist in the folder, the old file will not be overwrite for the
 safety of your model files.
-* -b: The mini-batch size. It is expected that the mini-batch size is a multiple of core_number.
 
 ## Test Model
 The above commands will cache the model in specified path(--checkpoint). Run this command will
@@ -41,16 +45,18 @@ use the model to do a validation.
 
 Example command
 ```
-java -cp dist/lib/bigdl-VERSION-jar-with-dependencies.jar \
+java -cp spark/dl/target/bigdl-VERSION-jar-with-dependencies-and-spark.jar \
 com.intel.analytics.bigdl.example.lenetLocal.Test \
 -f path_to_mnist_folder \
 --model ./model/model.iteration \
+-c core_number \
 -b batch_size
 ```
 In the above command
 * -f: where you put your MNIST data
 * --model: the model snapshot file
-* -b: The mini-batch size.
+* -c: The core number on local machine used for this testing. The default value is physical cores number. Get it through Runtime.getRuntime().availableProcessors() / 2
+* -b: The mini-batch size. It is expected that the mini-batch size is a multiple of core_number
 
 ## Predict with Model
 The above commands will use the model in specified path(--checkpoint)to do a prediction with given data.
@@ -60,8 +66,10 @@ Example command
 java -cp dist/lib/bigdl-VERSION-jar-with-dependencies.jar \
 com.intel.analytics.bigdl.example.lenetLocal.Predict \
 -f path_to_mnist_folder \
+-c core_number \
 --model ./model/model.iteration
 ```
 In the above command
 * -f: where you put your MNIST data
+* -c: The core number on local machine used for this prediction. The default value is physical cores number. Get it through Runtime.getRuntime().availableProcessors() / 2
 * --model: the model snapshot file
